+++ src/tools/Built-in/3_foc/foc.data.ts
@@ -21,7 +21,7 @@ export type TStatementElement = Readonly<{
 }>;
 
 /**
- * after initialization clear
+ * Generate .json indirect reference, no longer direct reference
  * FlowOfControl
  * FOC
  */